+++ Saobracaj/packages.config
@@ -11,7 +11,7 @@
   <package id="Microsoft.Report.Viewer" version="11.0.0.0" targetFramework="net461" />
   <package id="Microsoft.ReportingServices.ReportViewerControl.WebForms" version="150.1404.0" targetFramework="net461" />
   <package id="Microsoft.ReportViewer.Common" version="10.0.40219.1" targetFramework="net461" />
-  <package id="Microsoft.SqlServer.Types" version="14.0.314.76" targetFramework="net461" />
+  <package id="Microsoft.SqlServer.Types" version="14.0.1016.290" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net461" />
   <package id="OpenPop.NET" version="2.0.6.1120" targetFramework="net45" />
   <package id="Stub.System.Data.SQLite.Core.NetFramework" version="1.0.115.0" targetFramework="net461" />
@@ -66,7 +66,7 @@
   <package id="Syncfusion.Tools.Windows" version="18.4.0.34" targetFramework="net461" />
   <package id="Syncfusion.TreeMap.Windows" version="18.4.0.34" targetFramework="net461" />
   <package id="Syncfusion.XlsIO.WinForms" version="18.4.0.34" targetFramework="net461" />
-  <package id="System.Data.SqlClient" version="4.8.2" targetFramework="net461" />
+  <package id="System.Data.SqlClient" version="4.8.3" targetFramework="net461" />
   <package id="System.Data.SQLite" version="1.0.115.0" targetFramework="net461" />
   <package id="System.Data.SQLite.Core" version="1.0.115.0" targetFramework="net461" />
   <package id="System.Data.SQLite.EF6" version="1.0.115.0" targetFramework="net461" />
